+++ Public/js/modules/image.js
@@ -1,3 +1,112 @@
+class ImageLibrary {
+    constructor() {
+        this.selectedImage = false;
+        this.hasMore = false;
+        if (!window.linkcmsImages) {
+            window.linkcmsImages = {
+                images: [],
+                imageIds: [],
+                offset: 0,
+                total: false
+            };
+        }
+        this.imageStore = window.linkcmsImages;
+        this.buildElements();
+        this.loadFromLocal();
+        this.loadImages();
+        this.addEvents();
+
+    }
+
+    acceptUpload(e) {
+        if (this.uploadModal.uploader.uploadData && this.uploadModal.uploader.uploadData.type == 'success') {
+            let data = this.uploadModal.uploader.uploadData;
+            let image = new Image(data.content.image, this, true);
+            this.imageStore.imageIds.push(image.id);
+            this.imageStore.images.push(image);
+            this.selectImage(image.imageButton);
+            this.uploadModal = false;
+        } else {
+            this.uploadModal = false;
+        }
+    }
+
+    addEvents() {
+        let library = this;        
+        this.addButton.addEventListener('click', function(e) {
+            e.preventDefault();
+            library.uploadModal = new ImageUploadModal();
+            library.uploadModal.modalContainer.addEventListener('uploaded', library.acceptUpload.bind(library));
+        });
+    }
+
+    addImages() {
+        if (this.ajaxload.response.type == 'success') {
+            if (this.ajaxload.response.content.count) {
+                this.imageStore.total = this.ajaxload.response.content.count;
+            }
+            if (this.ajaxload.response.content.images.length > 0) {
+                this.imageStore.offset += this.ajaxload.response.content.images.length;
+                let Library = this;
+                this.ajaxload.response.content.images.forEach(function(imageData) {
+                    let image = new Image(imageData, Library);
+                    Library.imageStore.imageIds.push(image.id);
+                    Library.imageStore.images.push(image);
+                });
+                this.ajaxload = false;
+            }
+        }
+    }
+
+    buildElements() { 
+        this.label = new DomEl('label.sr-only');
+        this.label.innerText = 'This modal will allow you to select an image from the library. Use the add button to upload a new image, or select an existing one using the spacebar to select an image and enter to confirm it.';
+        this.container = new DomEl('div.library.thumbnailView');
+        this.labelContainer = new DomEl('div.sr-only');
+        this.addButton = new DomButton('addImage', 'file-upload', 'primary', ' Upload');
+        this.imageHolder = new DomEl('div.images.display-flex');
+        this.container.append(this.labelContainer);
+        this.container.append(this.label);
+        this.container.append(this.addButton);
+        this.container.append(this.imageHolder);
+    }
+
+    loadFromLocal() {
+        if (this.imageStore.images.length > 0) {
+            this.localLoad = true;
+            let Library = this;
+            this.imageStore.images.forEach(function(image) {
+                image.buildElement(Library);
+            });
+        }
+    }
+
+    loadImages() {
+        if (!this.imageStore.total || this.imageStore.images.length != this.imageStore.total) {
+            let url ='/api/image/load/' + this.imageStore.offset;
+            this.ajaxload = new Ajax(url, false, false, 'GET', true);
+            let LibraryObj = this;
+            this.ajaxload.eventEl.addEventListener('success', function() {
+                LibraryObj.addImages();
+            }); 
+        }
+    }
+
+    selectImage(imageButton) {
+        if (imageButton.classList.contains('selected')) {
+            imageButton.classList.remove('selected');
+            this.selectedImage = false;
+        } else {
+            let extant = this.imageHolder.querySelectorAll('.selected');
+            if (extant.length > 0) {
+                extant[0].classList.remove('selected');
+            }
+            this.selectedImage = imageButton;
+            this.selectedImage.classList.add('selected');
+        }
+    }
+}
+
 class ImageUploader {
     constructor() {
         this.required = ['altText','title'];
@@ -53,7 +162,6 @@ class ImageUploader {
                         this.required.forEach(function(requiredId) {
                             if (!uploader.fields[requiredId].value) {
                                 hasAllFields = false;
-                                console.log('falsified');
                             }
                         });
                         if (!hasAllFields) {
@@ -76,21 +184,22 @@ class ImageUploader {
                     }
                     this.form.style.display = 'none';
                     let bar = new ProgressBar(this.container);
-                    let upload = new Ajax(url, data, bar);
-                    window.upload = upload;
-                    upload.addEventListener('success', function(e) {
+                    this.upload = new Ajax(url, data, bar, 'POST', true);
+                    this.upload.eventEl.addEventListener('success', function(e) {
                         bar.track.classList.add('success');
-                        uploader.createImageEl(e.target.getAttribute('imageUrl'),e.target.getAttribute('altText'));
+                        uploader.uploadData = uploader.upload.response;
                         uploader.form.dispatchEvent(new Event('uploaded'));
                         bar.track.remove();
                         uploader.uploading = false;
+                        uploader.upload = false;
                     });
-                    upload.addEventListener('failure', function(e) {
+                    this.upload.eventEl.addEventListener('failure', function(e) {
                         if (uploader.uploading) {  
                             uploader.uploading = false;
                             bar.track.classList.add('failure');
                             bar.track.remove();
                             uploader.form.style.display = 'block';
+                            uploader.upload = false;
                         }
                     });
                 } else {
@@ -113,9 +222,11 @@ class ImageUploader {
         span.innerText = 'Click here to browse or drop the image you want to upload';
         let icon = new DomEl('i.fas.fa-file-image');
         this.label = new DomEl('label.imageUploader[for="uploader"][tabindex=0][title="Hit enter to browse for an image to upload"]');
-        this.preview = new DomEl('img.preview'); 
+        let previewDiv = new DomEl('div.preview');
+        this.preview = new DomEl('img.preview');
+        previewDiv.append(this.preview); 
         this.label.append(icon);
-        this.label.append(this.preview);
+        this.label.append(previewDiv);
         this.label.append(document.createElement('br'));
         this.label.append(span);
         this.label.append(this.input);
@@ -135,10 +246,6 @@ class ImageUploader {
         this.fields.caption = captionLabel.children[0];
     }
 
-    createImageEl(url, altText) {
-        this.imageEl = new DomEl('img[src=' + url + '][alt=' + altText + '].chosen');
-    }
-
     setContent(content) {
         if (content.image) {
             this.input.value = content.image;
@@ -149,5 +256,51 @@ class ImageUploader {
         if (content.altText) {
             this.fields.altText.value = content.altText;
         }
+    }
+}
+
+class Image {
+    constructor(imageData, library, addToTop=false) {
+        for (let [key,value] of Object.entries(imageData)) {
+            this[key] = value;
+        }
+        this.buildElement(library, addToTop);
+    }
+    
+    addEventListeners() {
+        let library = this.library;
+        this.imageButton.addEventListener('click', function(e) {
+            e.preventDefault(); 
+            library.selectImage(this);
+        });
+    }
+    
+    addToLibrary(addToTop) {
+        if (addToTop) {
+            this.library.imageHolder.insertBefore(this.imageButton, this.library.imageHolder.querySelectorAll('button')[0]);
+        } else {
+            this.library.imageHolder.append(this.imageButton);
+        }
+        this.library.labelContainer.append(this.label);
+    }
+    
+    buildElement(library, addToTop) {
+        this.library = library;
+        let labelId = 'image' + new Date().getMilliseconds();
+        this.label = new DomEl('label.sr-only#' + labelId);
+        this.label.innerText = 'Image title: "' + this.title + '". Image description: ' + this.altText;
+        let container = new DomEl('div.imageContainer');
+        let imageObj = new DomEl('img[src=' + this.imageUrl + ']');
+        let title = new DomEl('h3');
+        title.innerText = this.title;
+        this.imageButton = new DomEl('button.image.box.p-4[aria-describedby="' + labelId +']');
+        for (let [key,value] of Object.entries(this)) {
+            this.imageButton.dataset[key] = value;
+        };
+        container.append(imageObj);
+        this.imageButton.append(container);
+        this.imageButton.append(title);
+        this.addToLibrary(addToTop);
+        this.addEventListeners();
     }
 }